requirements.txt
aiofiles
aiohttp
emoji==1.6.0
python-dotenv
attrify
bs4
uvloop
dnspython
GitPython
gpytranslate
filetype
httpx[http2]
gtts
html2text
motor
opencv-python-headless
openai
Pillow==9.5.0
psutil
pykeyboard
pySmartDL
pyromod==1.5
pytz
pymediainfo
pymongo
pyrogram
python-dateutil
removebg
requests
telegraph
tgcrypto
unidecode
wget
youtube-search-python
yt-dlp
pydub
pytgcalls==3.0.0.dev24
search_engine_parser
